+++ Commands/Export-PipeScript.ps1
@@ -50,18 +50,28 @@ function Export-Pipescript {
 
             $buildFileTemplate = $buildFile.Template
             if ($buildFileTemplate -and $buildFile.PipeScriptType -ne 'Template') {
-                Invoke-PipeScript $buildFileTemplate.Source
+                try {
+                    Invoke-PipeScript $buildFileTemplate.Source
+                } catch {
+                    $ex = $_
+                    Write-Error -ErrorRecord $ex
+                }
                 $alreadyBuilt[$buildFileTemplate.Source] = $true
             }
 
-            $EventsFromThisBuild = Get-Event | 
+            $EventsFromThisBuild = Get-Event |
                 Where-Object TimeGenerated -gt $ThisBuildStartedAt |
-                Where-Object SourceIdentifier -Like 'PipeScript.*'
+                Where-Object SourceIdentifier -Like '*PipeScript*'
             
-            Invoke-PipeScript $buildFile.Source 
+            try {
+                Invoke-PipeScript $buildFile.Source
+            } catch {
+                $ex = $_
+                Write-Error -ErrorRecord $ex
+            }
+
             $alreadyBuilt[$buildFile.Source] = $true
         }
-
 
         $BuildTime = [DateTime]::Now - $buildStarted
         Write-Progress "Building PipeScripts [$FilesToBuildCount / $filesToBuildTotal]" "Finished In $($BuildTime) " -Completed -id $filesToBuildID