+++ Transpilers/Templates/Eiffel.Template.psx.ps1
@@ -1,38 +1,16 @@
 <#
 .SYNOPSIS
-    ADA Template Transpiler.
+    Eiffel Template Transpiler.
 .DESCRIPTION
-    Allows PipeScript to be used to generate ADA.
+    Allows PipeScript to be used to generate Eiffel.
     
-    Because ADA Scripts only allow single-line comments, this is done using a pair of comment markers.
+    Because Eiffel only allow single-line comments, this is done using a pair of comment markers.
 
     -- { or -- PipeScript{  begins a PipeScript block
 
     -- } or -- }PipeScript  ends a PipeScript block                
-.EXAMPLE
-    Invoke-PipeScript {
-        $AdaScript = '    
-    with Ada.Text_IO;
-
-    procedure Hello_World is
-    begin
-        -- {
-
-        Uncommented lines between these two points will be ignored
-
-        --  # Commented lines will become PipeScript / PowerShell.
-        -- param($message = "hello world")        
-        -- "Ada.Text_IO.Put_Line (`"$message`");"
-        -- }
-    end Hello_World;    
-    '
-    
-        [OutputFile('.\HelloWorld.ps1.adb')]$AdaScript
-    }
-
-    Invoke-PipeScript .\HelloWorld.ps1.adb
 #>
-[ValidatePattern('\.ad[bs]$')]
+[ValidatePattern('\.e$')]
 param(
 # The command information.  This will include the path to the file.
 [Parameter(Mandatory,ValueFromPipeline,ParameterSetName='TemplateFile')]