+++ Transpilers/Templates/Kusto.Template.psx.ps1
@@ -1,38 +1,20 @@
 <#
 .SYNOPSIS
-    ADA Template Transpiler.
+    Kusto Template Transpiler.
 .DESCRIPTION
-    Allows PipeScript to be used to generate ADA.
-    
-    Because ADA Scripts only allow single-line comments, this is done using a pair of comment markers.
+    Allows PipeScript to generate Kusto files.
 
-    -- { or -- PipeScript{  begins a PipeScript block
+    Multiline comments with /*{}*/ will be treated as blocks of PipeScript.
 
-    -- } or -- }PipeScript  ends a PipeScript block                
-.EXAMPLE
-    Invoke-PipeScript {
-        $AdaScript = '    
-    with Ada.Text_IO;
+    Multiline comments can be preceeded or followed by 'empty' syntax, which will be ignored.
 
-    procedure Hello_World is
-    begin
-        -- {
+    The Kusto Template Transpiler will consider the following syntax to be empty:
 
-        Uncommented lines between these two points will be ignored
-
-        --  # Commented lines will become PipeScript / PowerShell.
-        -- param($message = "hello world")        
-        -- "Ada.Text_IO.Put_Line (`"$message`");"
-        -- }
-    end Hello_World;    
-    '
-    
-        [OutputFile('.\HelloWorld.ps1.adb')]$AdaScript
-    }
-
-    Invoke-PipeScript .\HelloWorld.ps1.adb
+    * ```null```
+    * ```""```
+    * ```''```
 #>
-[ValidatePattern('\.ad[bs]$')]
+[ValidatePattern('\.kql$')]
 param(
 # The command information.  This will include the path to the file.
 [Parameter(Mandatory,ValueFromPipeline,ParameterSetName='TemplateFile')]
@@ -55,17 +37,20 @@ $ArgumentList
 
 begin {
     # We start off by declaring a number of regular expressions:
-    $startComment = '(?>(?<IsSingleLine>--)\s{0,}(?:PipeScript)?\s{0,}\{)'
-    $endComment   = '(?>--\s{0,}\}\s{0,}(?:PipeScript)?\s{0,})'        
-    $startRegex = "(?<PSStart>${startComment})"
-    $endRegex   = "(?<PSEnd>${endComment})"
+    $startComment = '/\*' # * Start Comments ```\*```
+    $endComment   = '\*/' # * End Comments   ```/*```
+    $Whitespace   = '[\s\n\r]{0,}'
+    # * IgnoredContext ```String.empty```, ```null```, blank strings and characters
+    $IgnoredContext = "(?<ignore>(?>$("null", '""', "''" -join '|'))\s{0,}){0,1}"
+    # * StartRegex     ```$IgnoredContext + $StartComment + '{' + $Whitespace```
+    $startRegex = "(?<PSStart>${IgnoredContext}${startComment}\{$Whitespace)"
+    # * EndRegex       ```$whitespace + '}' + $EndComment + $ignoredContext```
+    $endRegex   = "(?<PSEnd>$Whitespace\}${endComment}\s{0,}${IgnoredContext})"
 
     # Create a splat containing arguments to the core inline transpiler
     $Splat      = [Ordered]@{
         StartPattern  = $startRegex
         EndPattern    = $endRegex
-         # Using -LinePattern will skip any inline code not starting with --
-        LinePattern   = "^\s{0,}--\s{0,}"
     }
 }
 
@@ -79,6 +64,16 @@ process {
     
     if ($Parameter) { $splat.Parameter = $Parameter }
     if ($ArgumentList) { $splat.ArgumentList = $ArgumentList }
+
+    $splat.ForeachObject = {
+        $in = $_
+        if (($in -is [string]) -or 
+            ($in.GetType -and $in.GetType().IsPrimitive)) {
+            $in
+        } else {
+            "dynamic($(ConvertTo-Json -Depth 100 -InputObject $in))"
+        }
+    }
 
     # If we are being used within a keyword,
     if ($AsTemplateObject) {