+++ Transpilers/Templates/PS1XML.Template.psx.ps1
@@ -1,12 +1,59 @@
 <#
 .SYNOPSIS
-    XML Template Transpiler.
+    PS1XML Template Transpiler.
 .DESCRIPTION
-    Allows PipeScript to generate XML.
+    Allows PipeScript to generate PS1XML.
 
     Multiline comments blocks like this ```<!--{}-->``` will be treated as blocks of PipeScript.
+.EXAMPLE
+    $typesFile, $typeDefinition, $scriptMethod = Invoke-PipeScript {
+
+        types.ps1xml template '
+        <Types>
+            <!--{param([Alias("TypeDefinition")]$TypeDefinitions) $TypeDefinitions }-->
+        </Types>
+        '
+
+        typeDefinition.ps1xml template '
+        <Type>
+            <!--{param([Alias("PSTypeName")]$TypeName) "<Name>$($typename)</Name>" }-->
+            <!--{param([Alias("PSMembers","Member")]$Members) "<Members>$($members)</Members>" }-->
+        </Type>
+        '
+
+        scriptMethod.ps1xml template '
+        <ScriptMethod>
+            <!--{param([Alias("Name")]$MethodName) "<Name>$($MethodName)</Name>" }-->
+            <!--{param([ScriptBlock]$MethodDefinition) "<Script>$([Security.SecurityElement]::Escape("$MethodDefinition"))</Script>" }-->
+        </ScriptMethod>
+        '
+    }
+
+
+    $typesFile.Save("Test.types.ps1xml",
+        $typeDefinition.Evaluate(@{
+            TypeName='foobar'
+            Members = 
+                @($scriptMethod.Evaluate(
+                    @{
+                        MethodName = 'foo'
+                        MethodDefinition = {"foo"}
+                    }
+                ),$scriptMethod.Evaluate(
+                    @{
+                        MethodName = 'bar'
+                        MethodDefinition = {"bar"}
+                    }
+                ),$scriptMethod.Evaluate(
+                    @{
+                        MethodName = 'baz'
+                        MethodDefinition = {"baz"}
+                    }
+                ))
+        })
+    )
 #>
-[ValidatePattern('\.xml$')]
+[ValidatePattern('\.ps1xml$')]
 param(
 # The command information.  This will include the path to the file.
 [Parameter(Mandatory,ValueFromPipeline,ParameterSetName='TemplateFile')]